Add setting to disable Redis.
This introduces the concept of `WP_REDIS_DISABLED` to the `.env.example` file, and if it is defined, will disable Redis entirely.

The default value in the environment file is set to `true`, to disable it, while not having this defined at all will default to enabling it. This ensures we don't accidentally disable it if any sites wish to update Project Base at any point.

See #195