Add custom logpath to `WP_DEBUG_LOG`.
Although logging is only enabled if a development environment is enabled, these sites may also be publicly exposed on the internet.

Creating a log directory in the root of Project Base puts these logs outside the publicly available area of a website, removing the risk of potentially sharing sensitive information.

It also moves the registration of `WP_DEBUG_LOG` out of the environment specific files, and sets it on the application level in `application.php` for all environments. This ensures that if someone accidentally enables debug logging outside of development environments, it will always lead to logs ending up in our dedicated directory.