MU Plugin: Site Health quality assurance.
The Site Health feature in WordPress is great for indicating problems with a site, but depending on your setup, can also provide some false-positive results that are unexpected or unwanted.

This introduces the concept of a MU plugin that tackles this by removing tests with known false-positives relating to updates, plugins and themes, which are all managed via version control and Composer.

It also implements the `WP_ENVIRONMENT_TYPE` constant, that mirrors the existing `WP_ENV` constant. This ensures that plugins or themes that rely on WordPress' native environment types will get an expected result.