Composer: Implement `allow-plugins` and set a version for `dekode-theme`.
The smallest part of this change is moving the `dekode/dekode-theme` package from using the version constraint `@dev` to `~1.0.0`, this will avoid unnecessary lockfile-changes when working in branches and doing `composer update`.

Secondly, with Composer requiring any pluggable packages to be explicitly declared to be allowed to run, starting July 2022, this change preempts that by declaring them under `config.allow-plugins` starting now. This also works as an example of best practice for declarations going forward.