GitHub Actions: Update action runners.
Add artifact cleanups. Artifacts are reused through a workflow, but are also tied to the workflow, and not needed afterwards.

By clearing these out right after use, we lower our storage space use on GitHub to just a few minutes, allowing for many more concurrent projects before incurring additional charges.

The process can fail, but with artifacts configured to expire after 1 day (the lowest GitHub supports), we do not need the workflow to be treated as a failure just because the artifact is still there for a few hours.