GitHub Actions: Update action runners. (#450)
* GitHub Actions: Update action runners.

Switch to v3 of the cache action to prevent warnings about JS engines.

* GitHub Actions: Update action runners.

Use the new format for saving output from a job step.

* GitHub Actions: Update action runners.

Add artifact cleanups. Artifacts are reused through a workflow, but are also tied to the workflow, and not needed afterwards.

By clearing these out right after use, we lower our storage space use on GitHub to just a few minutes, allowing for many more concurrent projects before incurring additional charges.

The process can fail, but with artifacts configured to expire after 1 day (the lowest GitHub supports), we do not need the workflow to be treated as a failure just because the artifact is still there for a few hours.

* GitHub Actions: Update action runners.

Add artifact cleanups. Artifacts are reused through a workflow, but are also tied to the workflow, and not needed afterwards.

By clearing these out right after use, we lower our storage space use on GitHub to just a few minutes, allowing for many more concurrent projects before incurring additional charges.

Make the `TAG` attribute optional when making a manual deployment, to grab the latest code by default as intended.